+++ R/app.R
@@ -81,13 +81,8 @@ is_command_switch <- function(e) {
   typeof(switch_expr) == "character" || is_simple_assignment_call(switch_expr)
 }
 
-command_switch_help_on_missing <- function(switch_expr) {
-  if (is.character(switch_expr)) {
-    return(identical(switch_expr, ""))
-  }
-
-  stopifnot(is_simple_assignment_call(switch_expr))
-  is.null(switch_expr[[3L]])
+command_switch_help_on_missing <- function(anno) {
+  !isFALSE((anno %||% list())[["required"]])
 }
 
 .simple_call_syms <-
@@ -127,6 +122,7 @@ get_app_inputs <- function(app, exprs = app$exprs, pos = integer()) {
         stop("Only one app command switch() block allowed per expression level")
       }
       switch_expr <- e[[2L]]
+      switch_anno <- parse_expr_anno(getSrcLineNo(exprs[i]), lines, is_hashpipe)
       branches <- as.list(e)[-(1:2)]
       if (".val_pos_in_exprs" %in% names(branches)) {
         stop('command name ".val_pos_in_exprs" not permitted.')
@@ -149,7 +145,7 @@ get_app_inputs <- function(app, exprs = app$exprs, pos = integer()) {
       commands$.val_pos_in_exprs <-
         c(pos, i, 2L, if (is.call(switch_expr)) 3L)
       attr(commands, "help_on_missing_command") <-
-        command_switch_help_on_missing(switch_expr)
+        command_switch_help_on_missing(switch_anno)
 
       next
     }