+++ inst/examples/todo.R
@@ -11,7 +11,7 @@
 store <- ".todo.yml"
 
 switch(
-  command <- NULL,
+  command <- "",
 
   #| title: Display the todos
   #| description: Print the contents of the todo list.